# cmd/mb3server/api/openapi.yaml
openapi: 3.0.3
info:
  title: MassBank3 API
  version: "3.0"
servers:
- url: https://api.server.test/v1
paths:
  /records:
    get:
      operationId: getRecords
      parameters:
      - description: "Filter by a list of instrument type, which will be associated\
          \ by OR."
        explode: true
        in: query
        name: instrument_type
        required: false
        schema:
          $ref: '#/components/schemas/InstrumentType'
        style: form
      - description: Filter by SPLASH.
        explode: true
        in: query
        name: splash
        required: false
        schema:
          $ref: '#/components/schemas/Splash'
        style: form
      - description: "Filter by a List of MS types, which are associated by OR. If\
          \ none is given, all is assumed."
        explode: true
        in: query
        name: ms_type
        required: false
        schema:
          $ref: '#/components/schemas/MSType'
        style: form
      - description: "Filter by ion mode. If none is given, both is assumed."
        explode: true
        in: query
        name: ion-mode
        required: false
        schema:
          $ref: '#/components/schemas/IonMode'
        style: form
      - description: Filter by compound name
        explode: true
        in: query
        name: compound_name
        required: false
        schema:
          $ref: '#/components/schemas/CompoundName'
        style: form
      - description: Filter by exact mass.
        explode: true
        in: query
        name: exact_mass
        required: false
        schema:
          $ref: '#/components/schemas/ExactMass'
        style: form
      - description: "Tolerance for mass filters. Default: 0.3"
        explode: true
        in: query
        name: mass_tolerance
        required: false
        schema:
          $ref: '#/components/schemas/MassTolerance'
        style: form
      - description: Filter by formula.
        explode: true
        in: query
        name: formula
        required: false
        schema:
          $ref: '#/components/schemas/Formula'
        style: form
      - description: Filter by peaks.
        explode: true
        in: query
        name: peaks
        required: false
        schema:
          $ref: '#/components/schemas/Peaks'
        style: form
      - description: "Intensity for peak search, defaults to 100."
        explode: true
        in: query
        name: intensity
        required: false
        schema:
          $ref: '#/components/schemas/Intensity'
        style: form
      - description: Filter by peak difference
        explode: true
        in: query
        name: peak_differences
        required: false
        schema:
          $ref: '#/components/schemas/Peaks'
        style: form
      - description: Filter by peak list.
        explode: true
        in: query
        name: peak_list
        required: false
        schema:
          $ref: '#/components/schemas/PeakList'
        style: form
      - description: "Maximum number of results. Default: 20"
        explode: true
        in: query
        name: limit
        required: false
        schema:
          $ref: '#/components/schemas/Limit'
        style: form
      - explode: true
        in: query
        name: page
        required: false
        schema:
          $ref: '#/components/schemas/Page'
        style: form
      - description: "The Instensity cutoff for Peaklist search. Default: 5"
        explode: true
        in: query
        name: intensity_cutoff
        required: false
        schema:
          $ref: '#/components/schemas/IntensityCutoff'
        style: form
      - description: Filter by InChIKey. Complete or Partial
        explode: true
        in: query
        name: inchi_key
        required: false
        schema:
          $ref: '#/components/schemas/InChIKey'
        style: form
      - description: a contributor (institution)
        explode: true
        in: query
        name: contributor
        required: false
        schema:
          $ref: '#/components/schemas/Contributor'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
          description: A list of records
      summary: Get a list of records
  /search:
    post:
      operationId: searchSpectra
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Search'
        description: The search parameters
        required: true
      responses:
        "200":
          content:
            applicaton/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
          description: The search result
      summary: Search for Spectra
  /records/{accession}:
    get:
      operationId: getRecord
      parameters:
      - description: The accession id of the record
        explode: false
        in: path
        name: accession
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MBRecord'
          description: A MassBank record
      summary: Get a MassBank record
  /records/{accession}/svg:
    get:
      operationId: getSVG
      parameters:
      - description: The accession id of the record
        explode: false
        in: path
        name: accession
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            image/svg+xml:
              schema:
                type: string
          description: A compound SVG
      summary: The SVG image for an accession
...
  /filter/browse:
    get:
      operationId: getBrowseOptions
      parameters:
      - description: "Filter by a list of instrument type, which will be associated\
          \ by OR."
        explode: true
        in: query
        name: instrument_type
        required: false
        schema:
          $ref: '#/components/schemas/InstrumentType'
        style: form
      - description: "Filter by a List of MS types, which are associated by OR. If\
          \ none is given, all is assumed."
        explode: true
        in: query
        name: ms_type
        required: false
        schema:
          $ref: '#/components/schemas/MSType'
        style: form
      - description: "Filter by ion mode. If none is given, both is assumed."
        explode: true
        in: query
        name: ion-mode
        required: false
        schema:
          $ref: '#/components/schemas/IonMode'
        style: form
      - description: a contributor (institution)
        explode: true
        in: query
        name: contributor
        required: false
        schema:
          $ref: '#/components/schemas/Contributor'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowseOptions'
          description: All browse options
      summary: get browse options
...
components:
  schemas:
    InstrumentType:
      items:
        type: string
      type: array
    Splash:
      type: string
    MSType:
      items:
        enum:
        - MS
        - MS2
        - MS3
        - MS4
        type: string
      type: array
    IonMode:
      type: string
    CompoundName:
      type: string
    ExactMass:
      description: "A mass prefixed with \"and\" or \"or\" before a colon, e.g. \"\
        or:122.0\"."
      pattern: "(and|or):[0-9]+(\\.[0-9]*)?"
      type: string
    MassTolerance:
      default: 0.3
      format: double
      type: number
    Formula:
      description: "A chemical formula prefixed with \"and\" or \"or\" before a colon,\
        \ e.g. \"or:C6H7N5\"."
      pattern: "(and|or):(([A-Z][a-z]*[0-9]*)|\\*)*"
      type: string
    Peaks:
      items:
        description: "A m/z prefixed with \"and\" or \"or\" before a colon, e.g. \"\
          or:143.0\"."
        pattern: "(and|or):[0-9]+(\\.[0-9]*)?"
        type: string
      type: array
    Intensity:
      default: 100
      description: Releative intensity for peak search
      maximum: 100
      minimum: 1
      type: integer
    PeakList:
      items:
        description: "A tuple of m/z and intensity, e.g. (193.4,57)"
        pattern: "\\([0-9]+(\\.[0-9]*)?,[0-9]+\\)"
        type: string
      type: array
    Limit:
      default: 20
      maximum: 500
      minimum: 1
      type: integer
    Page:
      default: 1
      minimum: 1
      type: integer
    IntensityCutoff:
      default: 5
      maximum: 100
      minimum: 1
      type: integer
    InChIKey:
      type: string
    Contributor:
      items:
        description: The short version of the contributor name
        type: string
      type: array
    SearchResult:
      example:
        metadata:
          isomer_count: 0
          git_commit: git_commit
          compound_count: 0
          spectra_count: 0
          limit: 1
          page: 1
          version: version
          result_count: 0
          timestamp: timestamp
        data:
        - score:
            key: ""
          spectra:
          - id: id
            title: title
          - id: id
            title: title
          data: "{}"
          svg: svg
          synonyms:
          - synonyms
          - synonyms
          mass: 7.061401241503109
          name:
          - name
          - name
          formula: formula
        - score:
            key: ""
          spectra:
          - id: id
            title: title
          - id: id
            title: title
          data: "{}"
          svg: svg
          synonyms:
          - synonyms
          - synonyms
          mass: 7.061401241503109
          name:
          - name
          - name
          formula: formula
      properties:
        metadata:
          $ref: '#/components/schemas/Metadata'
        data:
          description: A list of compounds with associated spectra.
          items:
            $ref: '#/components/schemas/SearchResult_data_inner'
          type: array
      type: object
    Search:
      example:
        Filters:
          MsType:
          - MS
          - MS
          IonMode: POSITIVE
          Instrumenttype:
          - Instrumenttype
          - Instrumenttype
          Contributor:
          - Contributor
          - Contributor
        Parameters: "{}"
        Similarity:
          Parameters: "{}"
          Intensities:
          - 5.962133916683182
          - 5.962133916683182
          Mz:
          - 1.4658129805029452
          - 1.4658129805029452
        Compound: "{}"
        Spectra:
          PeakDifference:
          - 6.027456183070403
          - 6.027456183070403
          Peaks:
          - 0.8008281904610115
          - 0.8008281904610115
          Splash: Splash
      properties:
        Filters:
          $ref: '#/components/schemas/Search_Filters'
        Compound:
          type: object
        Spectra:
          $ref: '#/components/schemas/Search_Spectra'
        Parameters:
          type: object
        Similarity:
          $ref: '#/components/schemas/Search_Similarity'
      type: object
    MBRecord:
      example:
        date:
          created: 2000-01-23
          modified: 2000-01-23
          updated: 2000-01-23
        copyright: copyright
        comments:
        - null
        - null
        mass_spectrometry:
          focused_ion:
          - subtag: BASE_PEAK
            value: value
          - subtag: BASE_PEAK
            value: value
          data_processing:
          - subtag: COMMENT
            value: value
          - subtag: COMMENT
            value: value
        deprecated:
          date: 2000-01-23
          reason: reason
        project: project
        accession: accession
        peak:
          annotation:
            values:
            - - values
              - values
            - - values
              - values
            header:
            - header
            - header
          peak:
            values:
            - intensity: 5.962133916683182
              rel: 563
              mz: 1.4658129805029452
            - intensity: 5.962133916683182
              rel: 563
              mz: 1.4658129805029452
            header:
            - header
            - header
          splash: splash
          numPeak: 1
        title: title
        compound:
          names:
          - names
          - names
          inchi: inchi
          smiles: smiles
          classes:
          - classes
          - classes
          mass: 0.8008281904610115
          cdk_depict:
          - cdk_depict
          - cdk_depict
          link:
          - identifier: identifier
            database: CAS
          - identifier: identifier
            database: CAS
          formula: formula
        license: CC0
        species:
          lineage:
          - lineage
          - lineage
          name: name
          link:
          - identifier: identifier
            database: CAS
          - identifier: identifier
            database: CAS
          sample:
          - sample
          - sample
        publication: publication
        acquisition:
          general:
          - subtag: CONCENTRATION
            value: value
          - subtag: CONCENTRATION
            value: value
          ion_mobility:
          - subtag: CCS
            value: value
          - subtag: CCS
            value: value
          mass_spectrometry:
            ms_type: MS
            subtags:
            - subtag: MS_TYPE
              value: value
            - subtag: MS_TYPE
              value: value
            ion_mode: POSITIVE
          chromatography:
          - subtag: CARRIER_GAS
            value: value
          - subtag: CARRIER_GAS
            value: value
          instrument: instrument
          instrument_type: instrument_type
        authors:
        - name: name
          marc-relator: abr
        - name: name
          marc-relator: abr
      properties:
        accession:
          description: Record identifier
          pattern: "MSBNK-[A-Z0–9_]{1,32}-[A-Z0–9_]{1,64}"
          type: string
        deprecated:
          $ref: '#/components/schemas/MBRecord_deprecated'
        title:
          description: "Short title of the record. It is composed as 'compound.names[0];\
            \ acquisition.instrument_type; acquisition.mass_spectrometry.ms_type'"
          type: string
        date:
          $ref: '#/components/schemas/MBRecord_date'
        authors:
          items:
            $ref: '#/components/schemas/Authors_inner'
          type: array
        license:
          description: Creative Commons License or its compatible terms
          enum:
          - CC0
          - CC BY
          - CC BY-NC
          - CC BY-NC-SA
          - CC BY-SA
          type: string
        copyright:
          description: Copyright
          type: string
        publication:
          description: Bibliographic information of reference
          type: string
        project:
          description: Information on a related project
          type: string
        comments:
          description: Comments
          items:
            $ref: '#/components/schemas/MBRecord_comments_inner'
          type: array
        compound:
          $ref: '#/components/schemas/MBRecord_compound'
        species:
          $ref: '#/components/schemas/MBRecord_species'
        acquisition:
          $ref: '#/components/schemas/MBRecord_acquisition'
        mass_spectrometry:
          $ref: '#/components/schemas/MBRecord_mass_spectrometry'
        peak:
          $ref: '#/components/schemas/MBRecord_peak'
      required:
      - accession
      - authors
      - date
      - license
      - title
      type: object
...
    Metadata:
      example:
        isomer_count: 0
        git_commit: git_commit
        compound_count: 0
        spectra_count: 0
        limit: 1
        page: 1
        version: version
        result_count: 0
        timestamp: timestamp
      properties:
        version:
          type: string
        timestamp:
          type: string
        git_commit:
          type: string
        spectra_count:
          minimum: 0
          type: integer
        compound_count:
          minimum: 0
          type: integer
        isomer_count:
          minimum: 0
          type: integer
        result_count:
          minimum: 0
          type: integer
        page:
          minimum: 1
          type: integer
        limit:
          minimum: 1
          type: integer
      type: object
    BrowseOptions:
      example:
        metadata:
          isomer_count: 0
          git_commit: git_commit
          compound_count: 0
          spectra_count: 0
          limit: 1
          page: 1
          version: version
          result_count: 0
          timestamp: timestamp
        compound_start:
        - count: 0
          value: value
        - count: 0
          value: value
        contributor:
        - count: 0
          value: value
        - count: 0
          value: value
        ms_type:
        - count: 0
          value: value
        - count: 0
          value: value
        ion_mode:
        - count: 0
          value: value
        - count: 0
          value: value
        instrument_type:
        - count: 0
          value: value
        - count: 0
          value: value
      properties:
        metadata:
          $ref: '#/components/schemas/Metadata'
        contributor:
          items:
            $ref: '#/components/schemas/StringCount_inner'
          type: array
        instrument_type:
          items:
            $ref: '#/components/schemas/StringCount_inner'
          type: array
        ms_type:
          items:
            $ref: '#/components/schemas/StringCount_inner'
          type: array
        ion_mode:
          items:
            $ref: '#/components/schemas/StringCount_inner'
          type: array
        compound_start:
          items:
            $ref: '#/components/schemas/StringCount_inner'
          type: array
      type: object
...
    SearchResult_data_inner:
      description: Data for a single compound.
      example:
        score:
          key: ""
        spectra:
        - id: id
          title: title
        - id: id
          title: title
        data: "{}"
        svg: svg
        synonyms:
        - synonyms
        - synonyms
        mass: 7.061401241503109
        name:
        - name
        - name
        formula: formula
      properties:
        score:
          additionalProperties: true
          type: object
        data:
          type: object
        name:
          description: Compound names.
          items:
            type: string
          type: array
        synonyms:
          description: More compound names.
          items:
            type: string
          type: array
        formula:
          description: Formula of the compound.
          type: string
        mass:
          description: Exact Mass of the compound.
          format: double
          type: number
        svg:
          description: Smiles to generate structure.
          type: string
        spectra:
          description: A list of spectra for the compound
          items:
            $ref: '#/components/schemas/SearchResult_data_inner_spectra_inner'
          type: array
      type: object
    Search_Filters:
      example:
        MsType:
        - MS
        - MS
        IonMode: POSITIVE
        Instrumenttype:
        - Instrumenttype
        - Instrumenttype
        Contributor:
        - Contributor
        - Contributor
      properties:
        Instrumenttype:
          items:
            type: string
          type: array
        MsType:
          items:
            enum:
            - MS
            - MS1
            - MS2
            - MS3
            type: string
          type: array
        IonMode:
          enum:
          - POSITIVE
          - NEGATIVE
          type: string
        Contributor:
          items:
            type: string
          type: array
      type: object
    Search_Spectra:
      example:
        PeakDifference:
        - 6.027456183070403
        - 6.027456183070403
        Peaks:
        - 0.8008281904610115
        - 0.8008281904610115
        Splash: Splash
      properties:
        Splash:
          type: string
        Peaks:
          items:
            format: double
            type: number
          type: array
        PeakDifference:
          items:
            format: double
            type: number
          type: array
      type: object
    Search_Similarity:
      example:
        Parameters: "{}"
        Intensities:
        - 5.962133916683182
        - 5.962133916683182
        Mz:
        - 1.4658129805029452
        - 1.4658129805029452
      properties:
        Mz:
          items:
            format: double
            type: number
          type: array
        Intensities:
          items:
            format: double
            type: number
          type: array
        Parameters:
          type: object
      type: object
    MBRecord_deprecated:
      description: Mark the record as deprecated
      example:
        date: 2000-01-23
        reason: reason
      properties:
        date:
          description: Date of deprecation
          format: date
          type: string
        reason:
          description: reason for deprecation
          type: string
      type: object
    MBRecord_date:
      description: Date of creation or last modification of record
      example:
        created: 2000-01-23
        modified: 2000-01-23
        updated: 2000-01-23
      properties:
        updated:
          description: Date of last update to record
          format: date
          type: string
        created:
          description: Date of record creation
          format: date
          type: string
        modified:
          description: Date of last (automatic) modification to record
          format: date
          type: string
      type: object
...
    MBRecord_comments_inner:
      oneOf:
      - description: A pure comment string
        type: string
      - $ref: '#/components/schemas/MBRecord_comments_inner_oneOf'
...
    MBRecord_species:
      description: Information of Biological Sample
      example:
        lineage:
        - lineage
        - lineage
        name: name
        link:
        - identifier: identifier
          database: CAS
        - identifier: identifier
          database: CAS
        sample:
        - sample
        - sample
      properties:
        name:
          description: Scientific name of biological species
          type: string
        lineage:
          description: Lineage of species
          items:
            description: A lineage class item
            type: string
          type: array
        link:
          description: External database name with identifier
          items:
            $ref: '#/components/schemas/DatabaseObject'
          type: array
        sample:
          description: Information of sample preparation
          items:
            description: "Tissue or Cell, from which Sample was Prepared"
            type: string
          type: array
      type: object
...
    StringCount_inner:
      example:
        count: 0
        value: value
      properties:
        value:
          type: string
        count:
          minimum: 0
          type: integer
      type: object
    Authors_inner:
      description: Name of the author or affiliation
      example:
        name: name
        marc-relator: abr
      properties:
        name:
          type: string
        marc-relator:
          description: Marc relator to describe contribution of author or affiliation
          enum:
          - abr
          - acp
          - act
          - adi
          - adp
          - aft
          - anl
          - anm
          - ann
          - ant
          - ape
          - apl
          - app
          - aqt
          - arc
          - ard
          - arr
          - art
          - asg
          - asn
          - ato
          - att
          - auc
          - aud
          - aui
          - aus
          - aut
          - bdd
          - bjd
          - bkd
          - bkp
          - blw
          - bnd
          - bpd
          - brd
          - brl
          - bsl
          - cas
          - ccp
          - chr
          - clb
          - cli
          - cll
          - clr
          - clt
          - cmm
          - cmp
          - cmt
          - cnd
          - cng
          - cns
          - coe
          - col
          - com
          - con
          - cor
          - cos
          - cot
          - cou
          - cov
          - cpc
          - cpe
          - cph
          - cpl
          - cpt
          - cre
          - crp
          - crr
          - crt
          - csl
          - csp
          - cst
          - ctb
          - cte
          - ctg
          - ctr
          - cts
          - ctt
          - cur
          - cwt
          - dbp
          - dfd
          - dfe
          - dft
          - dgc
          - dgg
          - dgs
          - dis
          - dln
          - dnc
          - dnr
          - dpc
          - dpt
          - drm
          - drt
          - dsr
          - dst
          - dtc
          - dte
          - dtm
          - dto
          - dub
          - edc
          - edm
          - edt
          - egr
          - elg
          - elt
          - eng
          - enj
          - etr
          - evp
          - exp
          - fac
          - fds
          - fld
          - flm
          - fmd
          - fmk
          - fmo
          - fmp
          - fnd
          - fpy
          - frg
          - gis
          - grt
          - his
          - hnr
          - hst
          - ill
          - ilu
          - ins
          - inv
          - isb
          - itr
          - ive
          - ivr
          - jud
          - jug
          - lbr
          - lbt
          - ldr
          - led
          - lee
          - lel
          - len
          - let
          - lgd
          - lie
          - lil
          - lit
          - lsa
          - lse
          - lso
          - ltg
          - lyr
          - mcp
          - mdc
          - med
          - mfp
          - mfr
          - mod
          - mon
          - mrb
          - mrk
          - msd
          - mte
          - mtk
          - mus
          - nrt
          - opn
          - org
          - orm
          - osp
          - oth
          - own
          - pad
          - pan
          - pat
          - pbd
          - pbl
          - pdr
          - pfr
          - pht
          - plt
          - pma
          - pmn
          - pop
          - ppm
          - ppt
          - pra
          - prc
          - prd
          - pre
          - prf
          - prg
          - prm
          - prn
          - pro
          - prp
          - prs
          - prt
          - prv
          - pta
          - pte
          - ptf
          - pth
          - ptt
          - pup
          - rbr
          - rcd
          - rce
          - rcp
          - rdd
          - red
          - ren
          - res
          - rev
          - rpc
          - rps
          - rpt
          - rpy
          - rse
          - rsg
          - rsp
          - rsr
          - rst
          - rth
          - rtm
          - sad
          - sce
          - scl
          - scr
          - sds
          - sec
          - sgd
          - sgn
          - sht
          - sll
          - sng
          - spk
          - spn
          - spy
          - srv
          - std
          - stg
          - stl
          - stm
          - stn
          - str
          - tcd
          - tch
          - ths
          - tld
          - tlp
          - trc
          - trl
          - tyd
          - tyg
          - uvp
          - vac
          - vdg
          - voc
          - wac
          - wal
          - wam
          - wat
          - wdc
          - wde
          - win
          - wit
          - wpr
          - wst
          type: string
      required:
      - name
      type: object
...